@@ -1,5 +1,12 @@
+import logo from './logo.svg';
 import './App.css';
-import React from "react";
-const App = () => <PortfolioContainer />;
+import PortfolioContainer from './components/PortfolioContainer';
+function App() {
+  return (
+    <div className="App">
+      <PortfolioContainer />
+    </div>
+  );
+}
 
 export default App;
